// src/lib/database/models/image.model.ts
import { Document, model, models, Schema } from 'mongoose'

export type Image = Document & {
  title: string
  transformationType: string
  publicId: string
  secureURL: URL
  width?: number
  height?: number
  config?: object
  transformationURL?: URL
  aspectRatio?: string
  color?: string
  prompt?: string
  author?: {
    id: string
    firstName: string
    lastName: string
  }
  createdAt?: Date
  updatedAt?: Date
}

const ImageSchema = new Schema({
  title: { type: String, required: true },
  transformationType: { type: String, required: true },
  publicId: { type: String, required: true },
  secureURL: { type: String, required: true },
  width: { type: Number },
  height: { type: Number },
  config: { type: Object },
  transformationURL: { type: String },
  aspectRatio: { type: String },
  color: { type: String },
  prompt: { type: String },
  author: { type: Schema.Types.ObjectId, ref: 'User' },
  createdAt: { type: Date, default: Date.now },
  updatedAt: { type: Date, default: Date.now },
})

const Image = models?.Image || model('Image', ImageSchema)

export default Image